Don't disable 22/tcp on firewall
We're already switching what port SSH daemon operates under, so we
should be able to leave the firewall open on 22/tcp and just not have
anything listening on it without negatively affecting anything. Doing so
lets us run things like InSpec on our Packer image without having to
re-work our SSH connection midway through the build.

If we really want to be diligent, removing 22/tcp should be done at
provision-time (think user-scripts) so we block it at the firewall, on
the off-chance something starts listening on 22.