Installs Vault
We defer enabling Vault agent or Vault server to a decision made when
provisioning the node. To the end-user it will appear the same: some
service available via HTTP on port 8200 that appears to be Vault. The
implementation of it varies drastically, though.

If this is a Vault server, we want to allow any (trusted) server to
communicate against 8200. If it's a Vault agent, we want only localhost
traffic to communicate. This will need to be established at provisioning
time so there's no confusion.