Consul agent should not expose some endpoints publicly
Doing so, if it's not a server, would risk attaching the ACL token
assigned by the agent to anyone who makes an HTTP call on this port.
Normally not a problem if coming from localhost, but definitely is a
privilege escalation problem when coming from somewhere else on the
network.